training/requirements.txt
tensorflow==2.16.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
mlflow==2.17.0
PyYAML==6.0.2
joblib==1.4.2
scipy==1.11.4
prefect==2.19.9
requests==2.32.3
evidently==0.4.39
whylogs==1.4.3
boto3==1.34.160
